doc, refactor, and style.
