--- Sastrawi.sln
@@ -1,35 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 14
-VisualStudioVersion = 14.0.24720.0
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{888888A0-9F3D-457C-B088-3A5042F75D52}") = "Sastrawi", "src\Sastrawi\Sastrawi.pyproj", "{9A2DB5B1-A5B0-4B52-AA46-D2CE718CC5DD}"
-EndProject
-Project("{888888A0-9F3D-457C-B088-3A5042F75D52}") = "SastrawiTest", "tests\SastrawiTest.pyproj", "{69199BE5-44C5-45C3-8B82-62F14DA2B9F1}"
-EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{0302964A-E17E-468E-8365-21827A654692}"
-	ProjectSection(SolutionItems) = preProject
-		.coveragerc = .coveragerc
-		.gitignore = .gitignore
-		.travis.yml = .travis.yml
-		README.md = README.md
-		README.rst = README.rst
-		setup.cfg = setup.cfg
-		setup.py = setup.py
-	EndProjectSection
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{9A2DB5B1-A5B0-4B52-AA46-D2CE718CC5DD}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{9A2DB5B1-A5B0-4B52-AA46-D2CE718CC5DD}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{69199BE5-44C5-45C3-8B82-62F14DA2B9F1}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{69199BE5-44C5-45C3-8B82-62F14DA2B9F1}.Release|Any CPU.ActiveCfg = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-EndGlobal
