+++ tests/stemmer3_test.py
@@ -6,8 +6,8 @@ class Test_StemmerTest(unittest.TestCase):
         self.stemmer = Stemmer(stopwords = ('di', 'ke'))
 
     def test_removeStopWord(self):
-        self.assertEquals('pergi sekolah', self.stemmer.remove('pergi ke sekolah'))
-        self.assertEquals('makan rumah', self.stemmer.remove('makan di rumah'))
+        self.assertEqual('pergi sekolah', self.stemmer.remove_stopword('pergi ke sekolah'))
+        self.assertEqual('makan rumah', self.stemmer.remove_stopword('makan di rumah'))
 
 if __name__ == '__main__':
     unittest.main()
